Moved common data middleware into separate module
