+++ routes/common-data/index.js
@@ -1,0 +1,55 @@
+/*jshint esversion: 6 */
+(function() {
+  'use strict';
+
+  const _ = require('lodash');
+  const Common = require(__dirname + '/../common');
+
+  module.exports = (app, config, ModulesClass) => {
+
+    app.use(function(req, res, next) {
+      var modules = new ModulesClass(config);
+
+      modules
+        .menus.list()
+        .fragments.list()
+        .callback(function(data) {
+          var menus = data[0];
+          var fragments = data[1];
+          
+          _.keys(menus).forEach(menuKey => {
+            var menu = menus[menuKey];
+
+            menu.items = menu.items.map(item => {
+              if (item.url) {
+                item.url = Common.processLink(null, item.url);
+              }
+
+              return item;
+            });
+          });
+          
+          var fragmentMap = {};
+          fragments.forEach((fragment) => {
+            fragmentMap[fragment.slug] = fragment.contents;
+          });
+
+          req.kuntaApi = {
+            data: {
+              menus: menus,
+              fragmentMap: fragmentMap
+            }
+          };
+
+          next();
+        }, (err) => {
+          next({
+            status: 500,
+            error: err
+          });
+        });
+    });
+    
+  };
+
+}).call(this);
