+++ routes/index.js
@@ -7,49 +7,6 @@
   const Common = require(__dirname + '/common');
 
   module.exports = function(app, config, ModulesClass) {
-
-    app.use(function(req, res, next) {
-      var modules = new ModulesClass(config);
-
-      modules
-        .menus.list()
-        .fragments.list()
-        .callback(function(data) {
-          var menus = data[0];
-          var fragments = data[1];
-          
-          _.keys(menus).forEach(menuKey => {
-            var menu = menus[menuKey];
-
-            menu.items = menu.items.map(item => {
-              if (item.url) {
-                item.url = Common.processLink(null, item.url);
-              }
-
-              return item;
-            });
-          });
-          
-          var fragmentMap = {};
-          fragments.forEach((fragment) => {
-            fragmentMap[fragment.slug] = fragment.contents;
-          });
-
-          req.kuntaApi = {
-            data: {
-              menus: menus,
-              fragmentMap: fragmentMap
-            }
-          };
-
-          next();
-        }, (err) => {
-       	  next({
-       	    status: 500,
-       	    error: err
-       	  });
-        });
-    });
     
     app.get('/redirect/:type/:id', (req, res, next) => {
       var type = req.params.type;
@@ -78,6 +35,10 @@
         break;
       }
     });
+    
+    // Register common data middleware
+    
+    require(__dirname + '/common-data')(app, config, ModulesClass);
     
     // Register routes
